More debug, less delay.
