+++ src/tomato/daemon/__init__.py
@@ -1,1 +1,1 @@
-from .main import main_loop, tomato_job
+from .main import main_loop
