+++ src/tomato/daemon/main.py
@@ -1,13 +1,11 @@
-from importlib import metadata
 import psutil
-import argparse
 import os
 import subprocess
 import time
 import json
 import logging
 
-from ..drivers import driver_worker, driver_reset
+from ..drivers import driver_worker, driver_reset, tomato_job
 from .. import dbhandler
 
 
@@ -45,71 +43,6 @@ def _pipeline_ready_sample(ret: tuple, sample: dict) -> bool:
             return True
         else:
             return False
-
-
-def tomato_job() -> None:
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--version",
-        action="version",
-        version=f'%(prog)s version {metadata.version("tomato")}',
-    )
-    parser.add_argument(
-        "jobfile",
-        help="Path to a ketchup-processed payload json file.",
-        default=None,
-    )
-    args = parser.parse_args()
-
-    logfile = args.jobfile.replace(".json", ".log")
-
-    logging.basicConfig(
-        level=logging.DEBUG,
-        format="%(asctime)s:%(levelname)-8s:%(processName)s:%(message)s",
-        handlers=[logging.FileHandler(logfile, mode="a"), logging.StreamHandler()],
-    )
-    logger = logging.getLogger(__name__)
-
-    logger.info("attempting to load jobfile '%s'", args.jobfile)
-    with open(args.jobfile, "r") as infile:
-        jsdata = json.load(infile)
-
-    logger.debug("parsing data from jobfile")
-    settings = jsdata["settings"]
-    payload = jsdata["payload"]
-    pipeline = jsdata["pipeline"]
-    pip = pipeline["name"]
-    jobid = jsdata["jobid"]
-    queue = settings["queue"]
-    state = settings["state"]
-    pid = os.getpid()
-
-    logger.debug(f"assigning job '{jobid}' on pid '{pid}' into pipeline '{pip}'")
-    dbhandler.pipeline_assign_job(state["path"], pip, jobid, pid, type=state["type"])
-    dbhandler.job_set_status(queue["path"], "r", jobid, type=queue["type"])
-    dbhandler.job_set_time(queue["path"], "executed_at", jobid, type=queue["type"])
-
-    logger.info("handing off to 'driver_worker'")
-    logger.info("==============================")
-    ret = driver_worker(settings, pipeline, payload, jobid, logfile)
-
-    logger.info("==============================")
-    ready = payload.get("tomato", {}).get("unlock_when_done", False)
-    if ret is None:
-        logger.info("job finished successfully, setting status to 'c'")
-        dbhandler.job_set_status(queue["path"], "c", jobid, type=queue["type"])
-    else:
-        logger.info("job was terminated, setting status to 'cd'")
-        dbhandler.job_set_status(queue["path"], "cd", jobid, type=queue["type"])
-        logger.info("handing off to 'driver_reset'")
-        logger.info("==============================")
-        driver_reset(settings, pipeline)
-        logger.info("==============================")
-        ready = False
-
-    logger.debug(f"setting pipeline '{pip}' as '{'ready' if ready else 'not ready'}'")
-    dbhandler.pipeline_reset_job(state["path"], pip, ready, type=state["type"])
-    dbhandler.job_set_time(queue["path"], "completed_at", jobid, type=queue["type"])
 
 
 def main_loop(settings: dict, pipelines: dict) -> None:
@@ -160,8 +93,8 @@ def main_loop(settings: dict, pipelines: dict) -> None:
                         jpath = os.path.join(root, "jobdata.json")
                         with open(jpath, "w") as of:
                             json.dump(args, of, indent=1)
-                        cfs = subprocess.CREATE_NEW_PROCESS_GROUP
-                        cfs |= subprocess.CREATE_NO_WINDOW
+                        cfs = subprocess.CREATE_NO_WINDOW
+                        cfs |= subprocess.CREATE_NEW_PROCESS_GROUP
                         subprocess.Popen(
                             ["tomato_job", str(jpath)],
                             creationflags=cfs,
