+++ src/tomato/drivers/__init__.py
@@ -1,2 +1,2 @@
 from . import biologic, dummy
-from .driver_funcs import driver_api, driver_worker, driver_reset
+from .driver_funcs import driver_api, driver_worker, driver_reset, tomato_job
