+++ src/tomato/drivers/logger_funcs.py
@@ -20,8 +20,8 @@ def log_listener(queue, configurer, path):
         logger.handle(record)
 
 
-def log_worker_config(queue):
+def log_worker_config(queue, loglevel = logging.INFO):
     h = logging.handlers.QueueHandler(queue)
     root = logging.getLogger()
     root.addHandler(h)
-    root.setLevel(logging.DEBUG)
+    root.setLevel(loglevel)
